Move sidebar styles to less
