+++ src/ViewerInner.tsx
@@ -16,6 +16,7 @@ import useToggle from './hooks/useToggle';
 import ExitFullScreenIcon from './icons/ExitFullScreenIcon';
 import PageLayer from './layers/PageLayer';
 import { Layout } from './layouts/Layout';
+import Sidebar from './layouts/Sidebar';
 import { RenderToolbarSlot } from './layouts/ToolbarSlot';
 import DropArea from './open/DropArea';
 import { PageSize } from './PageSizeCalculator';
@@ -26,7 +27,6 @@ import PrintZone from './print/PrintZone';
 import Match from './search/Match';
 import ScrollMode from './ScrollMode';
 import SelectionMode from './SelectionMode';
-import Sidebar from './Sidebar';
 import Toolbar from './Toolbar';
 import getFileExt from './utils/fileExt';
 import { SpecialLevel } from './zoom/zoomingLevel';
